% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_admtab.R
\name{admtab_server}
\alias{admtab_server}
\title{server-modul for administrative tabeller i Muskelregisteret}
\usage{
admtab_server(id, RegData, SkjemaOversikt, user)
}
\description{
server-modul for administrative tabeller i Muskelregisteret
}
