% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_datadump.R
\name{datadump_server}
\alias{datadump_server}
\title{server-modul for datadump muskelregisteret}
\usage{
datadump_server(id, userRole, reshID, mainSession)
}
\description{
server-modul for datadump muskelregisteret
}
