% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hj_funk.R
\name{dateInput2}
\alias{dateInput2}
\title{Alternativ til dateInput med mulighet for bare år, måned og år}
\usage{
dateInput2(inputId, label, minview = "years", maxview = "decades", ...)
}
\description{
Alternativ til dateInput med mulighet for bare år, måned og år
}
