% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hj_funk.R
\name{dateRangeInput2}
\alias{dateRangeInput2}
\title{Alternativ til dateRangeInput med mulighet for bare år, måned og år}
\usage{
dateRangeInput2(inputId, label, minview = "years", maxview = "decades", ...)
}
\description{
Alternativ til dateRangeInput med mulighet for bare år, måned og år
}
