% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuskelFigFordeling.R
\name{FigFordeling}
\alias{FigFordeling}
\title{S' - Vertikale eller hor. søyler
L' - Linjer}
\usage{
FigFordeling(RegData, libkat, outfile = "", valgtVar,
  datoFra = "2000-01-01", datoTil = "2050-01-01", minald = 0,
  maxald = 120, erMann = "", plotType = "S", diagnoseSatt = "")
}
\description{
S' - Vertikale eller hor. søyler
L' - Linjer
}

