% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_fordeling_grvar.R
\name{fordeling_grvar_server}
\alias{fordeling_grvar_server}
\title{server-modul for grupperte fordelingsfigurer i NORNMD sin shiny-app på
Rapporteket}
\usage{
fordeling_grvar_server(id, RegData, reshID, ss)
}
\value{
Modul fordelingsfigur
}
\description{
Kun til bruk i Shiny
}
