% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_fordeling_grvar.R
\name{fordeling_grvar_ui}
\alias{fordeling_grvar_ui}
\title{ui-modul for grupperte fordelingsfigurer i NORNMD sin shiny-app på
Rapporteket}
\usage{
fordeling_grvar_ui(id)
}
\value{
Modul fordelingsfigur
}
\description{
Kun til bruk i Shiny
}
