% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_fordelingsfig.R
\name{fordelingsfig_server}
\alias{fordelingsfig_server}
\title{server-modul for fordelingsfigurer i NORNMD sin shiny-app på Rapporteket}
\usage{
fordelingsfig_server(id, RegData, reshID)
}
\value{
Modul fordelingsfigur
}
\description{
Kun til bruk i Shiny
}
