% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_kumandel.R
\name{kumulativAndel_server}
\alias{kumulativAndel_server}
\title{server-modul for kumulativ andel i NORNMD sin shiny-app på Rapporteket}
\usage{
kumulativAndel_server(id, RegData, reshID, ss)
}
\value{
Modul
}
\description{
Kun til bruk i Shiny
}
