% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_kumandel.R
\name{kumulativAndel_ui}
\alias{kumulativAndel_ui}
\title{ui-modul for kumulativ andel i NORNMD sin shiny-app på Rapporteket}
\usage{
kumulativAndel_ui(id)
}
\value{
Modul
}
\description{
Kun til bruk i Shiny
}
