% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_rapport.R
\name{lag_SMArapport}
\alias{lag_SMArapport}
\title{Generer SMArapport og returner filnavn og sti til fil.}
\usage{
lag_SMArapport(baseName, reshID = 0)
}
\description{
Generer SMArapport og returner filnavn og sti til fil.
}
