% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_medikamentforlop.R
\name{medikament_sma_server}
\alias{medikament_sma_server}
\title{server-modul for medikamenttabeller i Muskelregisteret}
\usage{
medikament_sma_server(id, SMAoversikt, user)
}
\description{
server-modul for medikamenttabeller i Muskelregisteret
}
