% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_medikamentforlop.R
\name{medikament_sma_ui}
\alias{medikament_sma_ui}
\title{ui-modul for medikamenttabeller i Muskelregisteret}
\usage{
medikament_sma_ui(id)
}
\description{
ui-modul for medikamenttabeller i Muskelregisteret
}
