% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muskelRenderRmd.R
\name{muskel_kjor_autorapport}
\alias{muskel_kjor_autorapport}
\title{Render documents from rmarkdown files at Rapporteket for autorapport,
either subscription (or dispatchment, soon)}
\usage{
muskel_kjor_autorapport(
  report = "SMArapport_abo_v2.Rmd",
  outputType = c("html_document", "html_fragment", "pdf_document"),
  abonnement = TRUE,
  reshID = 4001031
)
}
\description{
Render documents from rmarkdown files at Rapporteket for autorapport,
either subscription (or dispatchment, soon)
}
