% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muskelBeregnIndikator.R
\name{muskelBeregnIndikator}
\alias{muskelBeregnIndikator}
\title{Calculate Indicator for Muskel Data}
\usage{
muskelBeregnIndikator(RegData, ind_id)
}
\arguments{
\item{RegData}{A data frame containing the registry data to be used for the calculation.}

\item{ind_id}{An identifier for the indicator to be calculated.}
}
\value{
A data frame or other object containing the calculated indicator results.
}
\description{
This function calculates an indicator based on the provided data and indicator ID.
}
\examples{
# Example usage:
# result <- muskelBeregnIndikator(RegData, ind_id = 1)

}
