% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuskelFigAndeler.R
\name{MuskelFigAndeler}
\alias{MuskelFigAndeler}
\title{Lag søylediagram som viser andeler av ulike variabler}
\usage{
MuskelFigAndeler(
  RegData,
  valgtVar = "Alder",
  datoFra = "2000-01-01",
  datoTil = "2050-01-01",
  reshID = 0,
  diagnosegr = -1,
  minald = 0,
  maxald = 120,
  erMann = 99,
  outfile = "",
  forlop = 99,
  egenavd = 0,
  diagnose = -1,
  undergr = -1,
  undergr2 = -1,
  enhetsUtvalg = 0,
  avdod = "",
  gen_aarsak_paavist = -1,
  UtredningsaarFra = 1950,
  UtredningsaarTil = as.numeric(format(Sys.Date(), "\%Y")),
  inkl_tittel = T,
  debutAlderFra = 0,
  debutAlderTil = 90
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel skal plottes}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren). Tekststreng skrevet som 'YYYY-mm-dd'}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren). Tekststreng skrevet som 'YYYY-mm-dd'}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{diagnosegr}{Diagnosegruppe
1: Muskelsykdommer
2: Spinal muskelatrofi
3: Polynevropati}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{erMann}{kjønn
1: menn
0: kvinner
99: begge (Default)}

\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}

\item{forlop}{Forløpstype
1: Basisregistrering
2: Oppfølging
3: Annet}

\item{egenavd}{0: Registrert ved HF
1: Følges opp ved HF
2: Diagnostisert ved HF
3: Bostatt i fylke}

\item{diagnose}{diagnosegrupper}

\item{undergr}{undergruppe av diagnosegruppen}

\item{undergr2}{undergruppe to}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet (Default)
1: Egen enhet mot resten av landet
2: Egen enhet}

\item{avdod}{avdød.}

\item{UtredningsaarFra}{Utredningsår f.o.m}

\item{UtredningsaarTil}{Utredningsår t.o.m}

\item{inkl_tittel}{inkluder titel. kan være TRUE eller FALSE}

\item{debutAlderFra}{debutalder f.o.m}

\item{debutAlderTil}{debutalder t.o.m}
}
\value{
En figur med søylediagram av ønsket variabel
}
\description{
Denne funksjonen lager et søylediagram som viser andeler av valgt variabel
filtrert på de utvalg som er gjort.
}
