% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuskelHentData.R
\name{MuskelHentRegData}
\alias{MuskelHentRegData}
\title{Provide global dataframe for NoRGast}
\usage{
MuskelHentRegData()
}
\value{
RegData data frame
}
\description{
Provides NoRGast data from staging
}
