% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuskelFigAndeler.R
\name{MuskelPlotAndeler}
\alias{MuskelPlotAndeler}
\title{Plotdelen av MuskelFigAndeler tatt ut som egen funksjon}
\usage{
MuskelPlotAndeler(plotdata, outfile = "", fargepalett = "BlaaRapp")
}
\arguments{
\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}
}
\value{
Fordelingsfigur over valgt variabel
}
\description{
Beskrevet over
}
