% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muskelBeregnIndikator.R
\name{muskelPlotIndikator}
\alias{muskelPlotIndikator}
\title{Plot Indicator for Muskel Data}
\usage{
muskelPlotIndikator(indikatordata, graaUt = NA, lavDG = NA, outfile = "")
}
\arguments{
\item{indikatordata}{A data frame or tibble containing the indicator data to be plotted.}
}
\value{
A plot object representing the visualized indicator data.
}
\description{
This function generates a plot for the given indicator data related to muskel analysis.
}
\details{
The function is designed to visualize the indicator data in a meaningful way.
Ensure that the input data is preprocessed and formatted correctly before passing it to this function.
}
\examples{
\dontrun{
# Example usage:
muskelPlotIndikator(indikatordata = my_data)
}

}
