% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuskelPreprosess.R
\name{MuskelPreprosess}
\alias{MuskelPreprosess}
\title{Preprossesser data for Muskelregisterets rapporter}
\usage{
MuskelPreprosess(RegData)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}
}
\value{
Et preprossessert datasett
}
\description{
Denne rensker og definerer opp nødvendige variabler
}
\details{
Her kan detaljer skrives
}

