% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuskelPrepVar.R
\name{MuskelPrepVar}
\alias{MuskelPrepVar}
\title{Preparer variabler for plotting}
\usage{
MuskelPrepVar(RegData, valgtVar, inkl_tittel = T)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel skal plottes}

\item{inkl_tittel}{inkluder titel. kan være TRUE eller FALSE}
}
\value{
PrepData En liste med plotrelevante størrelser
}
\description{
Denne funksjonen grupperer og klargjør variabler for andelsplot
}
\details{
Her kan detaljer skrives
}
