% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muskelRenderRmd.R
\name{muskelRenderRmd2}
\alias{muskelRenderRmd2}
\title{Render documents from rmarkdown files at Rapporteket}
\usage{
muskelRenderRmd2(
  sourceFile,
  outputType = c("html_document", "html_fragment", "pdf_document"),
  params = list(),
  keep_tex_on_pdf = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sourceFile}{Character string providing the path to the rmarkdown
source file.}

\item{outputType}{Character string specifying the output format. Must be one
of \code{c("pdf_document", "html_document", "html_fragment")}.}

\item{params}{List of report parameters (as named values) to override the
corresponding entries under \emph{params} in the rmarkdown document yaml
header. Default is \code{NULL} in which case no parameters as defined in the
rmarkdown document will be overridden.}
}
\value{
Character string with path to the rendered file or, if
\code{outputType} is set to "html_fragment", a character string providing an
html fragment. Files are named according to \code{tempfile()} with an empty
pattern and with the extension according to \code{outputType}.
}
\description{
Function that renders documents at Rapporteket from rmarkdown source files.
Output formats may be (vanilla) HTML or PDF based on our own pandoc latex
template or fragments of html when the result is to be embedded in existing
web pages. Rmarkdown allow parameters to be part of report processing. Thus,
parameters that are specific to reports must be provided (as a named list)
when calling \code{muskelRenderRmd()}.
}
