% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuskelTabeller.R
\name{MuskelTabeller}
\alias{MuskelTabeller}
\title{Tabeller for Muskelregisteret}
\usage{
MuskelTabeller(RegData, datoFra, datoTil, minald, maxald, erMann, enhetsUtvalg)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren). Tekststreng skrevet som 'YYYY-mm-dd'}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren). Tekststreng skrevet som 'YYYY-mm-dd'}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{erMann}{kjønn
1: menn
0: kvinner
99: begge (Default)}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Default)
2: Egen enhet}
}
\value{
Tabeller til bruk i samlerapport eller annet
}
\description{
Denne funksjonen lager et utvalg av tabeller for muskelregisteret
}

