% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuskelTabellerForlopPas.R
\name{MuskelTabellerForlopspas}
\alias{MuskelTabellerForlopspas}
\title{Forløps of pasient tabeller for Muskelregisteret}
\usage{
MuskelTabellerForlopspas(
  RegDt = RegData,
  tidFra = "2008-01-01",
  tidTil = Sys.Date(),
  aldmin = 0,
  aldmax = 120,
  kjoen = 99,
  tidenh = "aar",
  frlType = NULL,
  avd = "Nei",
  IDType = "PasientID"
)
}
\arguments{
\item{RegDt}{en dataramme}

\item{tidfra}{}
}
\value{
tabelldata med antall pasienter/forløp per sykhus
per tidsenhet
}
\description{
Denne funksjonen tar inn registerdata lager tabeller
av antall pasienter/forløp per sykehus per tidsenhet
}
