% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuskelLibUtvalg.R
\name{MuskelUtvalg}
\alias{MuskelUtvalg}
\title{Gjør utvalg av dataene}
\usage{
MuskelUtvalg(RegData, datoFra, datoTil, minald, maxald, erMann, diagnoseSatt,
  diagnosegr, forlop, fargepalett = "BlaaRapp")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren). Tekststreng skrevet som 'YYYY-mm-dd'}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren). Tekststreng skrevet som 'YYYY-mm-dd'}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{erMann}{kjønn
1: menn
0: kvinner
99: begge (Default)}

\item{diagnoseSatt}{Hvor ble pasienten opprinnelig utredet}

\item{diagnosegr}{Diagnosegruppe
1: Muskelsykdommer
2: Spinal muskelatrofi
3: Polynevropati}

\item{forlop}{Forløpstype
1: Basisregistrering
2: Oppfølging
3: Annet}

\item{fargepalett}{Hvilken fargepalett skal brukes i figurer (Default: BlaaRapp)}
}
\value{
UtData En liste bestående av det filtrerte datasettet, utvalgstekst for figur og tekststreng som angir fargepalett
}
\description{
Denne funksjonen gjør utvalg av dataene og returnerer det filtrerte datasettet, utvalgsteksten
og fargepaletten for bruk i figuren
}

