% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuskelFigAndelStabel.R
\name{MuskelFigAndelStabel}
\alias{MuskelFigAndelStabel}
\title{Lag grupperte stabelplot for valgte variabler}
\usage{
MuskelFigAndelStabel(RegData, valgtVar, datoFra = "2000-01-01",
  datoTil = "2050-01-01", reshID, diagnosegr = "", minald = 0,
  maxald = 120, erMann = 99, outfile = "", diagnoseSatt = 99,
  forlop = 99, enhetsUtvalg = 1, preprosess = F, hentData = F,
  incl_N = F)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel skal plottes}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren). Tekststreng skrevet som 'YYYY-mm-dd'}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren). Tekststreng skrevet som 'YYYY-mm-dd'}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{diagnosegr}{Diagnosegruppe
1: Muskelsykdommer
2: Spinal muskelatrofi
3: Polynevropati}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{erMann}{kjønn
1: menn
0: kvinner
99: begge (Default)}

\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}

\item{diagnoseSatt}{Hvor ble pasienten opprinnelig utredet}

\item{forlop}{Forløpstype
1: Basisregistrering
2: Oppfølging
3: Annet}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Default)
2: Egen enhet}

\item{preprosess}{Preprosesser data
FALSE: Nei (Default)
TRUE: Ja}

\item{hentData}{Gjør spørring mot database
FALSE: Nei, RegData gis som input til funksjonen (Default)
TRUE: Ja}
}
\value{
PrepData En figur med stabelplot av ønsket variabel
}
\description{
Denne funksjonen lager grupperte stabelplot for valgte variabler
}
\details{
Her kan detaljer skrives
}
