% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hj_funk.R
\name{MuskelTabellerForlopspas}
\alias{MuskelTabellerForlopspas}
\title{Unik pasient/pasientforløps  tabeller for Muskelregisteret}
\usage{
MuskelTabellerForlopspas(
  RegData,
  tidFra = "2008-01-01",
  tidTil = Sys.Date(),
  aldmin = 0,
  aldmax = 120,
  kjoen = 99,
  tidenh = "aar",
  frlType = NULL,
  avd = "Nei",
  IDType = "PasientID"
)
}
\arguments{
\item{RegData}{dataramme fra muskelregisteret med variabler som
forløpsID og paientID}

\item{tidFra}{tidligste dato. Tekststreng skrevet som 'YYYY-mm-dd'}

\item{tidTil}{seineste dato. Tekststreng skrevet som 'YYYY-mm-dd'}

\item{aldmin}{Alder, f.o.m}

\item{aldmax}{Alder, t.o.m}

\item{kjoen}{0-kvinner, 1-menn, 99 - begge}

\item{tidenh}{tidsenhet. "aar" - år eller "maned" - måned}

\item{frlType}{Forløpstype.
kan være Basisregistrering, Oppfølging, NULL}

\item{avd}{Avdød."Ja", "Nei"}

\item{IDType}{"pasientID" eller "forlopsID"}
}
\value{
tabelldata med antall pasienter/forløp per sykhus
per tidsenhet
}
\description{
Denne funksjonen tar inn registerdata lager tabeller
av antall pasienter/pasientforløp per sykehus per tidsenhet
}
