% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuskelLibUtvalg.R
\name{MuskelUtvalg}
\alias{MuskelUtvalg}
\title{Gjør utvalg av dataene}
\usage{
MuskelUtvalg(
  RegData,
  datoFra,
  datoTil,
  minald,
  maxald,
  erMann,
  egenavd = 0,
  enhetsUtvalg,
  diagnosegr = -1,
  forlop,
  diagnose = "-1",
  undergr = -1,
  undergr2 = -1,
  avdod = "",
  fargepalett = "BlaaRapp",
  reshID,
  UtredningsaarFra = 1950,
  UtredningsaarTil = as.numeric(format(Sys.Date(), "\%Y")),
  debutAlderFra = 0,
  debutAlderTil = 90,
  gen_aarsak_paavist = -1
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren). Tekststreng skrevet som 'YYYY-mm-dd'}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren). Tekststreng skrevet som 'YYYY-mm-dd'}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{erMann}{kjønn
1: menn
0: kvinner
99: begge (Default)}

\item{egenavd}{0: Registrert ved HF
1: Følges opp ved HF
2: Diagnostisert ved HF
3: Bostatt i fylke}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet (Default)
1: Egen enhet mot resten av landet
2: Egen enhet}

\item{diagnosegr}{Diagnosegruppe
1: Muskelsykdommer
2: Spinal muskelatrofi
3: Polynevropati}

\item{forlop}{Forløpstype
1: Basisregistrering
2: Oppfølging
3: Annet}

\item{diagnose}{diagnosegrupper}

\item{undergr}{undergruppe av diagnosegruppen}

\item{undergr2}{undergruppe to}

\item{avdod}{avdød.}

\item{fargepalett}{Hvilken fargepalett skal brukes i figurer (Default: BlaaRapp)}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{UtredningsaarFra}{Utredningsår f.o.m}

\item{UtredningsaarTil}{Utredningsår t.o.m}

\item{debutAlderFra}{debutalder f.o.m}

\item{debutAlderTil}{debutalder t.o.m}
}
\value{
UtData En liste bestående av det filtrerte datasettet, utvalgstekst for figur og tekststreng som angir fargepalett
}
\description{
Denne funksjonen gjør utvalg av dataene og returnerer det filtrerte datasettet, utvalgsteksten
og fargepaletten for bruk i figuren
}
